+++ scripts/test.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+../node_modules/mocha/bin/mocha ../frete.spec.js --no-timeout
