--- test.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-./node_modules/mocha/bin/mocha frete.spec.js --no-timeout
